Show overlay when we are outside the solar system
